requirements.txt
python-telegram-bot==13.12
pymorphy2==0.9.1
rsa==4.9